Fix reference error windows not defined when a worker tries to import some reducer slice
